test(widget): migrate more widget tests to integration tests
Add tests for: Avatar, Alert, Accordion, Gauge, RadioGroup, Switch

- Avatar: render tests for small and medium sizes
- Alert: dismiss, handle_key, height, level colors/icons, render variants
- Accordion: selection, add/remove sections, render with border
- Gauge: set/get value, render all styles, helpers
- RadioGroup: selection, handle_key, render
- Switch: toggle, render, handle_key, checked alias

Tests accessing private fields remain inline in source files.
