+++ src/widget/gauge.rs
@@ -612,11 +612,12 @@ pub fn battery(level: f64) -> Gauge {
         .thresholds(0.5, 0.2)
 }
 
+// Most tests moved to tests/widget_tests.rs
+// Tests below access private fields and must stay inline
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::layout::Rect;
-    use crate::render::Buffer;
 
     #[test]
     fn test_gauge_new() {
@@ -691,61 +692,19 @@ mod tests {
     }
 
     #[test]
-    fn test_gauge_render_all_styles() {
-        let styles = [
-            GaugeStyle::Bar,
-            GaugeStyle::Battery,
-            GaugeStyle::Thermometer,
-            GaugeStyle::Arc,
-            GaugeStyle::Circle,
-            GaugeStyle::Vertical,
-            GaugeStyle::Segments,
-            GaugeStyle::Dots,
-        ];
-
-        for style in styles {
-            let mut buffer = Buffer::new(30, 5);
-            let area = Rect::new(0, 0, 30, 5);
-            let mut ctx = RenderContext::new(&mut buffer, area);
-
-            let g = Gauge::new().style(style).percent(50.0);
-            g.render(&mut ctx);
-        }
-    }
-
-    #[test]
-    fn test_gauge_with_title() {
-        let mut buffer = Buffer::new(30, 3);
-        let area = Rect::new(0, 0, 30, 3);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let g = Gauge::new().title("CPU Usage").percent(75.0);
-        g.render(&mut ctx);
-
-        assert_eq!(buffer.get(0, 0).unwrap().symbol, 'C');
-    }
-
-    #[test]
-    fn test_gauge_set_value() {
-        let mut g = Gauge::new();
-        g.set_value(0.8);
-        assert_eq!(g.get_value(), 0.8);
-    }
-
-    #[test]
-    fn test_gauge_helper() {
+    fn test_gauge_helper_value() {
         let g = gauge().percent(50.0);
         assert_eq!(g.value, 0.5);
     }
 
     #[test]
-    fn test_percentage_helper() {
+    fn test_percentage_helper_value() {
         let g = percentage(75.0);
         assert_eq!(g.value, 0.75);
     }
 
     #[test]
-    fn test_battery_helper() {
+    fn test_battery_helper_fields() {
         let g = battery(80.0);
         assert!(matches!(g.style, GaugeStyle::Battery));
         assert_eq!(g.value, 0.8);
