+++ src/widget/switch.rs
@@ -446,44 +446,12 @@ pub fn toggle(label: impl Into<String>) -> Switch {
 impl_styled_view!(Switch);
 impl_props_builders!(Switch);
 
+// Most tests moved to tests/widget_tests.rs
+// Tests below access private fields and must stay inline
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::layout::Rect;
-    use crate::render::Buffer;
-
-    #[test]
-    fn test_switch_new() {
-        let s = Switch::new();
-        assert!(!s.is_on());
-    }
-
-    #[test]
-    fn test_switch_on() {
-        let s = Switch::new().on(true);
-        assert!(s.is_on());
-    }
-
-    #[test]
-    fn test_switch_toggle() {
-        let mut s = Switch::new();
-        assert!(!s.is_on());
-
-        s.toggle();
-        assert!(s.is_on());
-
-        s.toggle();
-        assert!(!s.is_on());
-    }
-
-    #[test]
-    fn test_switch_disabled() {
-        let mut s = Switch::new().disabled(true);
-        assert!(!s.is_on());
-
-        s.toggle();
-        assert!(!s.is_on()); // Should not change when disabled
-    }
 
     #[test]
     fn test_switch_label() {
@@ -512,90 +480,14 @@ mod tests {
     }
 
     #[test]
-    fn test_switch_render_default() {
-        let mut buffer = Buffer::new(20, 1);
-        let area = Rect::new(0, 0, 20, 1);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let s = Switch::new().on(true);
-        s.render(&mut ctx);
-        // Smoke test
-    }
-
-    #[test]
-    fn test_switch_render_all_styles() {
-        let styles = [
-            SwitchStyle::Default,
-            SwitchStyle::IOS,
-            SwitchStyle::Material,
-            SwitchStyle::Text,
-            SwitchStyle::Emoji,
-            SwitchStyle::Block,
-        ];
-
-        for style in styles {
-            let mut buffer = Buffer::new(20, 1);
-            let area = Rect::new(0, 0, 20, 1);
-            let mut ctx = RenderContext::new(&mut buffer, area);
-
-            let s = Switch::new().style(style);
-            s.render(&mut ctx);
-        }
-    }
-
-    #[test]
-    fn test_switch_with_label() {
-        let mut buffer = Buffer::new(30, 1);
-        let area = Rect::new(0, 0, 30, 1);
-        let mut ctx = RenderContext::new(&mut buffer, area);
-
-        let s = Switch::new().label("Dark Mode");
-        s.render(&mut ctx);
-
-        assert_eq!(buffer.get(0, 0).unwrap().symbol, 'D');
-    }
-
-    #[test]
     fn test_switch_label_right() {
         let s = Switch::new().label("Test").label_right();
         assert!(!s.label_left);
     }
 
     #[test]
-    fn test_switch_handle_key() {
-        use crate::event::Key;
-
-        let mut s = Switch::new().focused(true);
-        assert!(!s.is_on());
-
-        assert!(s.handle_key(&Key::Enter));
-        assert!(s.is_on());
-
-        assert!(s.handle_key(&Key::Char(' ')));
-        assert!(!s.is_on());
-    }
-
-    #[test]
-    fn test_switch_helper() {
-        let s = switch().on(true);
-        assert!(s.is_on());
-    }
-
-    #[test]
     fn test_toggle_helper() {
         let s = toggle("Enable");
         assert_eq!(s.label, Some("Enable".to_string()));
-    }
-
-    #[test]
-    fn test_switch_checked_alias() {
-        // Test checked() is an alias for on()
-        let s = Switch::new().checked(true);
-        assert!(s.is_on());
-        assert!(s.is_checked());
-
-        let s = Switch::new().checked(false);
-        assert!(!s.is_on());
-        assert!(!s.is_checked());
     }
 }
